requirements.txt
aiogram==3.20.0.post0
openai==1.82.0
python-dotenv==1.1.0
